+++ src/chain/solana.rs
@@ -184,7 +184,9 @@ impl TryFrom<Network> for SolanaChain {
             Network::Scroll => Err(FacilitatorLocalError::UnsupportedNetwork(None)),
             Network::Robinhood => Err(FacilitatorLocalError::UnsupportedNetwork(None)),
             Network::RobinhoodTestnet => Err(FacilitatorLocalError::UnsupportedNetwork(None)),
-            Network::ArcTestnet => Err(FacilitatorLocalError::UnsupportedNetwork(None)),
+            Network::Arc | Network::ArcTestnet => {
+                Err(FacilitatorLocalError::UnsupportedNetwork(None))
+            }
             Network::Near => Err(FacilitatorLocalError::UnsupportedNetwork(None)),
             Network::NearTestnet => Err(FacilitatorLocalError::UnsupportedNetwork(None)),
             Network::Stellar => Err(FacilitatorLocalError::UnsupportedNetwork(None)),